# feat(performance): type-check program only once in `nargo test`